requirements.txt
# Core
httpx>=0.27
fastapi>=0.115
uvicorn>=0.30
google-generativeai

# LED strip (Pi only — install on Pi)
# adafruit-circuitpython-neopixel-spi
# adafruit-blinka

# OLED display (Pi only)
# luma.oled
# Pillow
